admin and user sign in and paypal api
added the paypal api and the files for each screens and I also added and checked the login as an admin as well as as a user logging in and made sure each link routed where I set it up to go. Also had lots of errors in code to go fix per the console warnings.